+++ purchase_delivery_split_date/__manifest__.py
@@ -1,13 +1,14 @@
 # Copyright 2014-2016 Numérigraphe SARL
 # Copyright 2017 ForgeFlow, S.L.
+# Copyright 2021 Jacques-Etienne Baudoux (BCIM) <je@bcim.be>
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
 
 {
     "name": "Purchase Delivery Split Date",
-    "version": "13.0.1.0.5",
+    "version": "13.0.1.1.0",
     "summary": "Allows Purchase Order you confirm to generate one Incoming "
     "Shipment for each expected date indicated in the Purchase Order Lines",
-    "author": "Numerigraphe, ForgeFlow, Camptocamp, Odoo Community Association (OCA)",
+    "author": "Numerigraphe, ForgeFlow, Camptocamp, BCIM, Odoo Community Association (OCA)",
     "website": "https://github.com/OCA/purchase-workflow",
     "category": "Purchase Management",
     "license": "AGPL-3",